Add Distribution.Version.Internal
